ci: wire the radar Python sidecar into the gate
detect.sh reports python=true and the setup action already installs uv on
that basis, but taskfile.yml only ever included repo/go/node. So radar/ was
outside the contract entirely: app.py and decode.py were never linted, and
radar/tests/test_decode.py was run by nothing -- a test file sitting in the
tree that no gate had ever executed.

The standard says the taskfile wires one entry per language present, so this
was a real deviation rather than a rough edge. Adds .taskfiles/python.yml and
radar/ruff.toml, and includes python with dir: radar.

ONE DEVIATION FROM THE TEMPLATE, in the test target. The template runs
`uv run --with pytest pytest`, which assumes a pyproject.toml for uv to
resolve the project's dependencies from. This sidecar has only a pinned
requirements.txt, so that form yields an environment holding pytest and
nothing else -- and the tests import decode, which imports matplotlib, numpy
and pyart. Collection dies on ModuleNotFoundError, which is the exact trap
the template's own comment warns about, reached by a different route.
--with-requirements installs the pinned set, so the tests run against the
versions radar actually ships. The template form is kept as the fallback for
a sidecar with no requirements.txt.

Three ruff findings, all absorbed rather than suppressed:

  I001    decode.py import block unsorted
  RUF046  decode.py int(round(float(x))) x2 -- round() already returns int

The I001 fix was checked rather than trusted, because the import order there
is load-bearing: matplotlib.use("Agg") must run before matplotlib.pyplot is
imported or the headless backend is not selected. The block ruff sorts starts
below that call, so `import matplotlib` and the use("Agg") line are untouched
and pyplot stays after it. Verified by reading the result, not by assuming
the fixer is safe.

Formatter also rewrapped two long calls (line-length 100).

Verified: full `task ci` exits 0 across all 19 stages. radar's 2 tests now
actually run (~40s including the dependency install). Mutation-checked rather
than trusted -- perturbing _parse_band, the function RUF046 touched, fails
test_decode_nids_to_isobands, so that path is genuinely covered and not
incidentally green.

+++ radar/decode.py
@@ -9,12 +9,11 @@ os.environ.setdefault("MPLBACKEND", "Agg")
 import matplotlib
 
 matplotlib.use("Agg")
+import geojsoncontour
 import matplotlib.pyplot as plt
 import numpy as np
 import pyart
 from pyart.io.nexrad_level3 import NEXRADLevel3File
-
-import geojsoncontour
 
 # Isoband steps over the NWS reflectivity scale (dBZ), 5-dBZ bands.
 _DBZ_LEVELS = list(range(5, 80, 5))
@@ -62,7 +61,11 @@ def _read_site_code(path: str) -> str:
     """Parse the 3-char site code from the NIDS text header's AWIPS ID (e.g. 'N0BTLX' -> 'TLX')."""
     nfile = NEXRADLevel3File(path)
     try:
-        lines = [line for line in nfile.text_header.decode("ascii", "replace").splitlines() if line.strip()]
+        lines = [
+            line
+            for line in nfile.text_header.decode("ascii", "replace").splitlines()
+            if line.strip()
+        ]
         awips_id = lines[-1].strip()
         return awips_id[3:]
     finally:
@@ -83,7 +86,9 @@ def _contour_reflectivity(radar) -> tuple[dict[str, Any], list[float]]:
     fig, ax = plt.subplots()
     try:
         contour_set = ax.tricontourf(lon, lat, values, levels=_DBZ_LEVELS, extend="neither")
-        feature_collection = geojsoncontour.contourf_to_geojson(contour_set, ndigits=5, serialize=False)
+        feature_collection = geojsoncontour.contourf_to_geojson(
+            contour_set, ndigits=5, serialize=False
+        )
     finally:
         plt.close(fig)
 
@@ -98,4 +103,6 @@ def _contour_reflectivity(radar) -> tuple[dict[str, Any], list[float]]:
 def _parse_band(title: str) -> tuple[int, int]:
     """Parse geojsoncontour's 'LOW.00-HIGH.00 ' title into integer 5-dBZ band bounds."""
     low_str, high_str = title.strip().split("-")
-    return int(round(float(low_str))), int(round(float(high_str)))
+    # round() on a float already returns an int in Python 3, so the int() wrapper
+    # this used to carry was a no-op (RUF046).
+    return round(float(low_str)), round(float(high_str))
